+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 ## Unreleased
 
-（なし）
+- **Breaking**: 最小サポートをbeta52に引き上げ
+- **Breaking**: `FilterProcVideo::get_font` を削除し、`EditSection::get_font` に移動
+- **Breaking**: `ScriptModuleReturnValue`からCloneを削除
 
 ### デモプラグイン
 
-（なし）
+- regex-moduleを追加
 
 ## [0.34.0](https://github.com/sevenc-nanashi/aviutl2-rs/releases/tag/0.34.0)